# openapi-GW.yml
openapi: 3.0.0
info:
  title: API Geoespaciales
  description: Conjunto de APIs con datos de cartografía, socioeconómicos, y más.
  version: "1.0"
  contact: {}
servers: []

# --- NUEVA SECCIÓN ---
# Aquí aplicamos la seguridad a TODA la API.
# Le decimos que todos los endpoints requieren la llave 'api_key_header'.
security:
  - api_key_header: []

paths:
  /colonias/por-id:
    get:
      summary: Obtiene una colonia por sus claves geoestadísticas.
      operationId: ColoniasController_getByIds
      parameters:
        - name: cve_ent
          required: true
          in: query
          schema:
            example: 108
            type: number
        - name: cve_mun
          required: true
          in: query
          schema:
            example: 13002
            type: number
        - name: cve_loc
          required: true
          in: query
          schema:
            example: 216
            type: number     
      responses:
        "200":
          description: "Respuesta exitosa"
      tags:
        - Catálogos Cartografía
      # Aquí le decimos a este endpoint específico a dónde ir.
      x-google-backend:
        # ¡MUY IMPORTANTE! Usa la URL que te da Cloud Run, no tu dominio personalizado.
        address: https://nahu-d3yityfp4a-uc.a.run.app
  
  /colonias/por-coordenadas:
    get:
      summary: Busca colonias dentro de un radio a partir de coordenadas.
      operationId: ColoniasController_getByCoords
      parameters:
        - name: lat
          required: true
          in: query
          schema:
            example: 21.880449
            type: number
        - name: lng
          required: true
          in: query
          schema:
            example: -102.296179
            type: number
        - name: radio
          required: false
          in: query
          description: Radio de búsqueda en metros
          schema:
            example: 1000
            type: number
        # ... (tus parámetros se quedan igual)
      responses:
        "200":
          description: "Respuesta exitosa"
      tags:
        - Catálogos Cartografía
      # --- NUEVA SECCIÓN ---
      x-google-backend:
        # ¡MUY IMPORTANTE! Usa la URL que te da Cloud Run, no tu dominio personalizado.
        address: https://nahu-d3yityfp4a-uc.a.run.app

# La sección de tags principales que ya tenías está bien.
tags:
  - name: Catálogos Cartografía
    description: ""
  - name: Indicadores Socioeconómicos
    description: ""
  - name: Comercio
    description: ""
  - name: Vivienda
    description: ""
# Aquí definimos cómo funciona nuestra seguridad.
components:
  securitySchemes:
    # Este es el nombre interno de nuestra regla de seguridad.
    api_key_header:
      type: apiKey
      # Le decimos que la llave vendrá en un encabezado (header).
      in: header
      # Le decimos que el encabezado se llamará 'x-api-key'.
      name: x-api-key